requirements.txt
firebase-admin==1.0.0
psutil
google-api-python-client
boto
simplejson